+++ backend/src/routes/dashboardRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const dashboardController = require('../controllers/dashboardController');
+
+//Post for User
+router.get('/', dashboardController.dashboardCheck);
+
+module.exports = router;