+++ backend/src/server.js
@@ -1,3 +1,4 @@
+//Required for Use
 const express = require('express');
 const cors = require('cors');
 require('dotenv').config();
@@ -8,31 +9,14 @@ const port = process.env.PORT || 5000;
 app.use(cors());
 app.use(express.json());
 
-// Test Route for Setup
-/*
-app.get('/api/v1', (req, res) => {
-    res.json({message: 'Server is running'});
-});
-
-// Test Route for DB
-app.get('/api/v1/test-db', async (req, res) => {
-    console.log(process.env.DB_HOST);
-    let connection;
-    try {
-        connection = await pool.getConnection();
-        const rows = await connection.query('SELECT 1 + 1 AS solution');
-        res.json({message: 'DB is connected', solution: rows[0].solution});
-    } catch (error) {
-        res.status(500).json({message: 'DB connection failed', error: error.message});
-    } finally {
-        if (connection) connection.release();
-    }
-});
-*/
-
+//Get the Routes
 const userRoutes = require('./routes/usersRoutes');
-app.use('/api/users', userRoutes);
+const dashboardRoutes = require('./routes/dashboardRoutes');
 
+app.use('/api/users', userRoutes);
+app.use('/api/dashboard', dashboardRoutes);
+
+//Where the app is ran
 app.listen(port, () => {
     console.log(`Server is running on http://localhost:${port}`);
 });