+++ documentation/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adhers to [Semantic Versioning](https://semver.org/)
 - Backend now connects to the database.
 - Added an API to create a user.
 - Added an API to sign in as a user.
+- Added an API to get dashboard access.
 
 ### Changed